- add a series of tests to check the behaviour of ''updateQuality()'' according to the business rules in the requirements.
There weren't any tests before.
We add these tests before doing any refactoring, so that we can be sure that we don't break anything with our refactorings.

We had to adapt some tests to limit them to 100 years in the past.
Because the code contains a bug when the ''sellIn'' field has a negative overflow error.
We avoid this error by limiting the domain to 100 years.

Also added the requirements as javadoc on the ''updateQuality()'' method. We have an IDE that pops-up javadoc when floating over it, so let's use it.