- add conjured items that reduce in quality twice as fast.
Reworked the ''updateItems()'' method to account for conjured ''Item''s.
Reworked so that the nested if's are flattened out, reducing the cyclic complexity.
This makes it easier to add the new type of Item, and even later when still other types would need to be added.
Separated out an ''ItemUpdater'' interface, that can have different implementations, one for each type of ''Item'' with a different behaviour.

At this stage, the conjured items work, so technically, mission accomplished.
There are however further refactorings that can be done.
I definitely do not agree with the code ownership remark of the corner-goblin. :-)
There should not exist ownership of code, if the code belongs to the Gilded Rose company.
So we should be allowed to change it.
Only for code external to the company, can it be that we can't change the code.

I will see if there are further refactorings needed. If I have time I will take a look.